make items swipeable and make the bottom sheet expandable

+++ app/src/main/java/com/example/pipi/feature_login/domain/model/LoginResponse.kt
@@ -4,6 +4,6 @@ package com.example.pipi.feature_login.domain.model
 data class LoginResponse(
     val `data`: LoginResponseData,
     val message: String,
-    val status: Int,
+    val status: Double,
     val success: Boolean
 )
